resolved form issue

+++ app/api/investment_routes.py
@@ -19,7 +19,7 @@ def create_investment():
     new_investment = Investment(
         portfolio_id=int(data.get('portfolioId')),
         stock_id=data.get('stockId'),
-        num_shares=int(data.get('numShares'))
+        num_shares=float(data.get('numShares'))
     )
 
     db.session.add(new_investment)
@@ -45,7 +45,7 @@ def edit_investment(investmentId):
         return jsonify({'message': 'Investment not found'}), 404
 
     # Update the investment with new data
-    investment.num_shares = int(data.get('numShares'))
+    investment.num_shares = float(data.get('numShares'))
     investment.stock_id = data.get('stockId')
     investment.portfolio_id = int(data.get('portfolioId'))
